hash tree: more on verification, must check zeroes
when testing truncation with the following commands:

./merkle write -r 0 353 test.orig test.hash
cp test.orig test.trunc
./merkle truncate -r 0 350 test.trunc test.hash
./merkle write -r 350 353 test.orig test.hash

./merkle verify test.orig test.hash
./merkle verify test.trunc test.hash

i noticed that the resulting test.hash file would verify successfully
against both test.orig as well as test.trunc.  the test.trunc
verification is expected to fail, because the hashes for removed blocks
should contain all zeroes, resulting in a different parent hash.

the issue here is that verification is computing its expected parent
hashes from the hash file itself, and this version doesn't have zeroes
for those children.

the fix is for verifier::visit_node() to check which children are
outside of the valid block range, and check those hashes for zeroes

this valid range checking required modifications to visitor::visit(),
which needs to know the highest valid block instead of calculating an
upper bound based on the maximum depth

Signed-off-by: Casey Bodley <casey@linuxbox.com>